NLP/requirements.txt
nltk==3.5
python_Levenshtein==0.12.0
